requirements.txt
numpy==1.14.3
opencv-python==3.4.1
Pillow==5.1.0
